server/pyproject.toml
[tool.poetry]
name = "assignment"
version = "0.1.0"
description = ""
authors = ["LeeJiChan <jiclee@cuneo.co.kr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.6"
djangorestframework = "^3.14.0"
django-filter = "^23.3"
pytest = "^7.4.3"
psycopg2-binary = "^2.9.9"
django-rest-swagger = "^2.2.0"
drf-yasg = "^1.21.7"
django-cors-headers = "^4.3.0"
environ = "^1.0"
drf-spectacular = "^0.26.5"
ruff = "^0.1.3"
pre-commit = "^3.5.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.1.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
